Renamed json_parser, and attempted to fix jargame paths and imports.
