+++ .gitignore
@@ -4,5 +4,8 @@
 /src/core/__pycache__/
 /src/snowman/__pycache__/
 /src/__pycache__/
-write-noun-sentence.code-workspace
 /src/whack_a_mole/assets/audio/gtts/*.mp3
+write-noun-sentence.code-workspace
+src/jar_bingo/__pycache__/
+src/jar_bingo/ver_1.0/
+src/jar_bingo/tests.py
