+++ src/jar_bingo/main.py
@@ -1,4 +1,12 @@
+import os 
+import inspect
+import sys
+currentdir = os.path.dirname(os.path.abspath(inspect.getfile(inspect.currentframe())))
+parentdir = os.path.dirname(currentdir)
+sys.path.insert(0, parentdir)
+
+from src.jar_bingo.game import *
 # Run the game
 
 if __name__ == "__main__":
-    jar_bingo_game.main()
+    main()
