+++ src/jar_bingo/tests.py
@@ -2,7 +2,7 @@ from src.core.json_response_parser import *
 
 
 string_data = '[{"name": "Alice", "age": 30}, "name": "Bob", "age": 25},]'
-parsed_data = parse_dict_list(string_data)
+parsed_data = parse_json_response(string_data)
 print(parsed_data)
 for d in parsed_data:
     if d.get("name"):
