+++ src/jar_bingo/ver_1.0/LLM_ver1.py
@@ -219,7 +219,7 @@ def get_questions(model, _lvl_prep, _sentence_count, _correct_example, _with_jar
     generated_response = model.generate_text(prompt=prompt_input)  # guardrails=False
     print("generated_response", generated_response)
     # Remove spaces before "[" and after "]"
-    processed_response = parse_dict_list(generated_response)
+    processed_response = parse_json_response(generated_response)
     print("processed_response:", processed_response)
     # try:
     #     json_response = json.loads(processed_response)
